LUA - Ahwassa Weapon
The Ahwassa weapon has some control mechanics that keep the user from interacting with it, immediately after dropping it's bomb.  The intention was to prevent the user from instantly stopping the unit, and 'flinging' the bomb - and intentionally avoiding flying anywhere near the intended target.

This was a crude mechanic, and often just stalled the unit in the vicinity.

We have replaced this mechanic with something quite a bit more natural, that will temporarily impair the units ability to turn, without removing the ability to control it.

As a result it is now possible to target a unit, and plot the movement away (with Shift key).   This wasn't previously possible.