LUA - Weapon and DefaultWeapon logic
Having discovered some targeting issues related to missile weapons (ammo based) that would cause them to 'stall' and never fire again, unless cleared, we really needed to dig thru and better understand this process.

The problem, in a nutshell, these weapons when given a ground target - would never realize they had fired - and thus the target would remain, after firing - and any additional targets would never be fired upon.

In order to properly clear a target, after firing, we need to have the weapon be disabled - very briefly (2 ticks) in the WeaponPackingState, which all weapons now pass thru, even if they have no corresponding animation.   This delay impacts all CountedProjectile (ammo based) weapons.

Along the way, we were able to clarify some of the weapon process even more, refining the CanWeaponFire() function to resolve ammo weapons.

Re-added the 'FirstShot' flag to make it backwardly compatible with mod weapons that might be using it (BrewLAN Hades PD).

Essentially, rather than having all kinds of shortcuts and bypasses (the original code) - we now have almost all weapons pass thru ALL of the states in the process of firing - for visibility, and just as importantly, for those custom weapons that may wish to overload a particular stage of the process.   This permits the AI DEBUG tools to be used to fully explore the behavior of any weapon, in terms of it's firing behavior, it's timings, and it's general process.