UNITS - Ahwassa
When targeting ground, the Ahwassa will consider an attack complete after 1 drop (from 3), allowing other commands to be processed immediately.

A reduction in the firing arc (32 >> 28) and a reduction in the bomb drop range (90 >> 70), combine with the changes to the bomb drop mechanics of the weapon, to provide a much more natural, and uninterrupted, flight envelope after a bomb drop.