UNITS - UEF T2 Shield ship
This unit had long been living outside the shield formula in many ways - size, regen and recharge rates were all grossly overstated.

I wanted to maintain most of the shield size, without pushing the consumption up too far - and in the end - this ship lives somewhere between a mobile shield (with it's regen penalties) and a static shield (with it's ability to rapidly recharge).

As anticipated, consumption goes well up - (250 >> 625).
Shield Size ( 72 >> 62 )
Shield Regen ( 100 >> 68 )
Shield Recharge Time ( 22 >> 28 )
Shield Drain Recharge Time ( 5 >> 8 )

The mass, required to sustain these values, has increased the mass cost of the unit ( 848 >> 950 )
Unit HP has dropped ( 900 >> 550 ).

The upshot of these changes are clear in how you should perceive it - it's essentially a static shield building, mounted on a floating platform.   It's basically a mobile shield, but gets some of the benefits of a static shield.