LOUD_MISC - Integrated Storage
Reviewed the unit cap costs of both T2 and T3 Mass Fabs.

The UI was not properly reflecting the 0.8 cap cost of the upgraded T2 fab, and was reporting zero.   It was never zero - and - the 0.8 cap cost was something we put in, many many years ago, to encourage the use of what was a largely unused unit at that time.

Now that's its purpose, and value, are largely recognized, we'll put it at a hard ONE, so that the UI can report it fully - and we have updated the descriptions accordingly.