LUA - AI Naval base selection
For the longest time, picking a naval base position has been simply about finding the closest empty point - and that's not always such a good idea.

This new process will evaluate the threat map, and attempt to pick the safest AND closest position - which adds a lot more variability to the process.