LUA - AI TML behavior
With fixes to the target sequence for both nukes and TMLS, it's clear that these needed some review - not only to better understand if they are being effective - but to better signal that a unit is able to fire (and not stuck like previously).    We now check the flag to confirm this status.