BO ACU - Overcharge weapons
The OC weapon has a nasty behavior - it will disable the standard weapon entirely - and only re-enable it when the OC itself is disabled.

This was likely done, originally, to insure that the OC weapon has aim control - but that's not really necessary since it's aim control has a higher precedence.

I have, to allow more extensive testing, removed this disabling.