LUA - AI UnitsGreaterAtLocationInRange
This function normally takes a fixed value for the 'max' range parameter.   We have modified this function to utilize the base location's RallyPointRadius, if the range is NOT provided.

This will shape the function to adapt to the base it's being called from, which will be useful in restricting the 'reclaim' behavior, of the AI, in accordance to the base he is working from.

Please note:  The RallyPointRadius range is NOT wide enough to prevent another player from infringing upon the 'Perimeter' ring of a base.

The Perimeter, and Picket rings are the two outermost 'rings' of a base layout - and lie beyond the 'Staging' ring (which is where the rally points are located).