LUA - AI Defense building
A better transition to T3 defense structures involves having better conditions, to assess the need, and a reclaim strategy that provides the real estate for the MAIN base to utilize more T3 structures in the base plan

Builders for both the normal engineers, and the ACU are included.

The AI base templates have been expanded to include a large portion of the T2 PD locations as part of the T3 choices.