LUA - AI Transport Utilities - Escorts
While assembling transports for a lift mission, the AI will now also assign both a scout - and up to 4 fighters/transport - from the available pool.

At present, the fighters will simply 'guard' their respective transport - but the intention is to use the 'MovePlatoon' function to handle the fighters as a squad within the platoon, and range ahead of the transports, rather than essentially following.