LOUD_MISC - MetalWorld
In addition to the recent changes to unit cap cost for MetalWorld extractors, there are two additional changes as follows;

The unit cap bonus for SACU is reduced from 3 to 1.

The energy maintenance of extractors is doubled.

Both of these changes are intended to put some additional pressure on the unit cap considerations, when using this mod.   The energy requirement should attenuate the rapid buildup somewhat, while the SACU change will limit overall unit cap growth in this  particular high-mass scenario.