LUA - Beam weapon fix
Users were noting additional damage from beam weapons - and when we use the Weapon debugs, we could see that beam weapons were creating an additional 'free' collision event - at the moment of beam creation.

This appears to be related to the 'instahit' feature, and beam weapons were relying upon the BeamCollisionDelay value, showing up in a obscure function, to trigger this.

The original, vanilla code would reference the entire weapon blueprint, every time, and then if the value was ZERO or less (which I've never seen), it would create this free 'hit'.   Well, to say the least, there are a lot of beams with 0 beam collision delay values.

The was later compounded by LOUD, when we tidied up the data stream, and passed the value directly into the beam - to save the repetitive loading - but it wasn't being read - so every beam was getting this 'free' hit - for the last two months or so.

All tidied up now - the repetitive read is gone, and the data is now passing properly.