PROJECTILES - Sera Torpedo
While working on the Sera Torpedo launchers, I discovered that they are NOT entering the water properly.

The weapon muzzle is actually out of the water, on all 3 tiers, but pointed slightly down.   However, the target tracking almost immediately sends the torpedo right at the target - and - if that target is at, or near, water level, the torpedo is, in some cases actually AIRBORNE - and can actually collide with other projectiles - bad, very bad.

You can easily observe this when testing the weapon out by targeting ground.

To remedy this, we introduce some code that checks if the torpedo is being created in the water, or not.   If NOT, then the tracking is turned off - and the torpedo will fire in the direction of the muzzles - which in this case, is down, into the water.

A second bit of code, executes when the torpedo executes the OnEnterWater() event, which, after a tiny delay, re-engages the target tracking - sending the torpedo towards it's target - but now underwater.