UNITS - T1 Torpedo Launchers
A full pass on all T1 Torpedo Launchers

In general, Torpedo launchers will take on a greater value in naval defense.   Overall, a significant boost to DPS ( 120 >> 181 ), but with a lower rate of fire (much more potent torps).

The lifetime of the torpedo is slightly reduced (to reduce projectile counts and particle counts).

The targeting bones, and the collision boxes have been adjusted as well.  The Sera launcher was scaled up to be closer in size to the other factions.