UNITS - T2 Torpedo Launchers
As with the T1 launchers, a significant step up in DPS ( 240 >> 300 ) - but with a drop in Rate of fire.   I've also added a brief STUN effect against T1 and T2 land units, and T1 Naval.

Firing tolerance and turret rotation speed, have both been trimmed, to make flanking against them more effective, now that they are more lethal.

All T2 launchers have a veteran buff that increases WaterVision radius.   In general, torpedo launchers have the best overall WaterVision radius of any units in the game.

Torpedo lifetime has been slightly reduced, to better align with the maximum range.  Both the Aeon and Sera systems have slightly greater range ( 64 versus 60 ).  For the Aeon, it's a factional buff, but for the Sera, it's a result of the elevated firing system.   This change will keep the total number of torpedoes, in the water, at any one time, in good check.

The Sera T2 launcher differs from the others with a smaller volley of torpedoes (3 instead of 4 and only 290 DPS), but the addition of an anti-torpedo system.

All T2 Launchers are both radar and sonar stealth, but once detected, remain so, as with all structures.