AI - Priority Functions
One of the methods we use to avoid a lot of repetitive conditional testing is to have a job priority drop below 100.   I often do this with a simple priority function.

Lately, I've been working with moving a good number of common, static, conditions away from being conditionals, and just being a priority function.

This CAN have beneficial impacts, first - the conditional can be entirely removed, and if it's a unique one, it reduces the size of the conditional table (which is processed repeatedly over the course of the game).

Second, it can reduce the size of the result tables, saving some memory, in those cases where a very large table can drop, in one or more orders of allocation.   This can have minor savings in lookups as well.