LUA - Loudutilities
Refactoring

Some new work on the Attack Planner.   The Attack Plan Monitor, will now review each stage of the plan - and mark the current stage of its progress.

If a stage is used, by an Allied AI - the current plan will be closed - and a new one will be built.   This will keep Allied AI from stepping on each other.

Fixed a ranging issue with stage building - that was causing some very poor stage selection.