LUA - Loiterpositions
SetLoiterPosition() is intended to find positions to stage air formations from, and search for targets.   This process was searching for the least difficult positions, rather than the most valuable.

Removed the commands that cleared the commands and issued the assist order.   This allows us to use the function to interrogate if the position has changed, while a formation is in combat.   This may result in the formation staying 'on target' as the best choice drifts towards, or away, from a certain target.

I've added in logging so that the results can be viewed but it's not a part of the AI Debug.

The clear commands, and the subsequent Assist orders, are now placed just after the call is made, in the appropriate places.