LUA - AirForceAILOUD
This is the fighter based behavior.   It makes use of the loiter process to position air formations nearer to potentially attractive targets.

Target selection, and attack tracking, are based on the loiterposition, so, if the fight takes the formation beyond it's search range (based on the loiter) it will disengage.

By adding the SetLoiterPosition() into the attack tracking - we sometimes will see the loiter position shift towards the current target - keeping the fight going.

Also, noted that IssueClearCommands were not functioning properly - since the Platoon Entity is not valid in that context, it must be a table of the units in the platoon.

Added in some primitive logging, but like the LoiterPosition logging, it's not a part of the AI Debug Tools as of yet.