UNITS - Mobile Bomb
The detonation process, of the mobile bomb, left a lot to be desired - so we've gone the route of using the same mechanics as the mines - blow up when near an enemy.

The difference here is that the mobile bombs can be put on Hold Fire - and then detonated manually if desired.

Some other minor descriptive additions, and some refactoring included here