LUA - Factory Bias
We introduced the Naval factory bias some months ago, to better detect a rush naval effort on the part of the human player - it accomplishes this by assuming that a player with naval factories, that are building something, will indeed register a competitive naval ratio.    It's been quite successful at thwarting the usually reliable sub rush to close the naval markers.

We're now adding this ability to his detection of air rush - and this works a little bit differently in that it will alter LOUD's production focus - and that, in turn, will influence how, and where, he 'loiters' his air power.