LUA - Prebuilt Units/Outnumbered
In conjunction with the other outnumbered mechanics, the prebuilt units (mass and T1 energy) deserved an examination.

When an AI is outnumbered by 2 to 1 or more, he'll get twice as many pre-built units (4 mex, 4 T1 Pgen) as his more numerous opponents.

The self-upgrade threads, for these units, get a staggered start, to insure they don't all attempt to upgrade on the same game tick.