UNITS/BO - T3 Subs
The torpedo defense on the Yathsou was over-reactive - as the re-firing delay was too low.   This was compounded by the presence of two (2) anti-torp emitters.

The delay moves from 7 to 10 on Yathsou.
The delay moves from 6 to 8 on Leviathan.

As a class, the E cost has been increased by 5% and water vision normalized at 40.