BO for LOUD - UEF ACU
The UEF ACU was first alerted to me due to it's missile firing issue.   I have re-installed the short delay, after firing - and fixed the weapon processing cycle to address the issues it was having.

Cleaned up some layers, for various weapons, that just can't happen for ACU - and removed some old comments.