LUA - DefaultWeapons, Weapon
Using some semaphores to track targeting, and, in conjunction with a weapon being enabled, or not, run sub-functions like OnLostTarget()

Moved the state changes, when enabling/disabling a weapon, into Weapon

Improved weapon logging with these new semaphores, to insure that we only see logging that is actually doing something to the weapon - so, for example, we won't see DeadState weapons processing OnLostTarget()

One process that has seen refinement is the processing of valid targets for the current layer.   Normally, this only runs on enabled weapons - so if a weapon was upgraded, after the unit had moved to another layer - the weapon would not fire.    This validation now occurs whenever a weapon is enabled - and when the unit transitions to another layer.