LUA - Defaultweapons
Default beam weapons, such as those used in mines and/or Sudeths melee weapons - don't have all the values expected from a beam weapon, so I have added some default guard values when those are not present.

You'll notice the need for a RackBone to be specified.

Also fixed a function that was expecting bp to already be present when it's not.