LUA - Lobby work
Trying to make some sense of how the Mod Manager works in conjunction with the lobby - and if it's possible to update the options without needing to reset the lobby.

Unfortunately, while we can get an AI Mod to load/unload - the lobby options appear to be loaded in advance - and no attempt at making them reload and refresh, seems to be available.

I did re-sequence some of the existing options, and moved relevant ones into the correct sub-groups - those options that previously had drop down selections, but are now free form edit fields, I was able to remove the additional, but unused data.