LUA - defaultweapons
The RackSalvoCharge state is responsible for processing any remaining economy events, and any salvo charge delays and visual effects that take place prior to actually firing.   When it completes, it moves the weapon, by default to the RackSalvoFireReady state.

The state had a specified OnFire() which immediately forced the weapon into RackSavloFireReady state, bypassing any remaining economy events, charge delays or other pre-firing visual affects.

The side effect of this was noted on the UEF T3 Rocket Pod (which has an unpacking animation that is triggered here) - as a result of the OnFire() event - the unpack animation was not being triggered - and the pod would fire from below ground.