LUA - Default weapons
During WeaponPackingState() the OnGotTarget() event was being trapped - and would circumvent the process - and send the weapon directly to WeaponUnpackingState().   This was problematic - and would cause the weapon to go inert - since OnGotTarget() and OnFire() are not trapped in that state.  By properly completing the pack - and going to IdleState() the problem is cured.