LUA - Weapons states, disabling and ammo weapons
The handling of ammo-based weapons is tricky - both SML and TML are marked as 'manual fire' weapons, which implies one shot at a time behavior.    These weapons also have a 'repack' process, which, should, temporarily disable the weapon, which should intentionally remove the current target - and any sequential targets that may have been queued up.

The Packing State used to do this - but we uncovered that other non-ammo but 'packing' weapons (Gatling bots) use the packing process to do simple cosmetic things - and if the weapon is disabled - the unit will want to 'grapple' with it's target instead of remaining at range.

A minor guard for CalcBallAcc, which is trying to use a deprecated blueprint value

To further complicate this - the process of enabling a weapon always returned it to 'Idle' State - which is undesirable behavior.

I had to dis-entangle that, which was an artifact of weapon creation - so that enabling/disabling a weapon has no implied state change.