Provide assistance with excessive UI scaling
When the main menu is displayed, a check is made to ensure that all
buttons are visible on the screen. If this is not the case, a dialog is
shown with the option to reset the scaling factor to the default value
of 100%. This dialog only appears once per game start.