LUA - Fix to prevent errors from missing blueprints
When loading map based units and wrecks, the authors sometimes will use a mod unit - or a civilian unit that we have deprecated - and this would throw a long series of error messages in the log.

By inserting a simple category check, during this process, we can sidestep those errors.