LUA - AIBehaviors - AirForceAILOUD
Continued refinement of the fighter behavior, most importantly, discovered the crash bug being caused by platoon merging taking place when fighters where still in the Attack squad (they'd lose connection to the source platoon).  I needed to toggle the merge and DistressResponse semaphore correctly so that it was only enabled, when the platoon was not engaging targets.

More refinement to the searchradius and it's relation to AirRatio and the platoon size.

While loitering, the cycle time will be lengthened by the number of units in the platoon.

Prior to RTB, all platoon orders are cleared.