Back to the roots
Unfortunately, I seem to be the only one who mourns the animmations of
the main menu from the original game. By chance I stumbled across the
code, most of which was just disabled. For what was missing, including
the timings, there was the original code. So here it is: The animations
of the main menu are back! There is now a well-hidden option at the
bottom of the interface options to activate the animations and set their
speed. Of course, the animations are disabled by default. Maybe nobody
will notice and therefore won't complain about it :)