LUA - Distress Response tuning
Enhanced the logging so it can be more easily analyzed.

Fixed some logic and response issues, and an issue with it not always terminating properly.  This meant that some platoons could either stall, or only respond once.

DR can now respond to platoon distress from other brains - this is a big feature.

DR can now track the location of the calling platoon, and check it's position (rather than the position of the original distress call)

Reaction time check rate is now throttled by distance from the distress location, but overall repeat rate is much higher as we get closer, than it had been previously