LUA - SelfUpgrade
This key process had a nasty flaw in it that was allowing upgradeable structures to 'drop out' of the process.

Extensive tuning work to review the impact that the bug was having on what I thought was happening

I now use a callback to trigger the next follow-on upgrade, eliminating a large chunk of repeated code.