LUA - Air unit Refit process
The refit process needed some careful study to better understand why some units would get 'stuck' to the airpad, or the air would go 'stupid' and stop taking air units.

Revised logging to better study the process.

In it's natural behavior - everything is just fine - the problems creep in when we try to 'force' and unload - by doing so, the clamps become 'busy' and unable to take another attach - and it was permanent.   We have removed ALL forced detaches - and replaced it with a simple monitor that waits for the natural detach to happen.