LUA - AIBrain metrics
Utilizing the intel data, we have developed a couple of new metrics, that refine some of the baseline data.  Production ratio, which gives us a relative value about just how much production capacity, and the nature of it, that an enemy might have, will help to guide the AI in determining the very general direction of an enemies production.

At present, this is purely indicative, and doesn't yet impact our own production.   The idea would be to use this to stimulate either building additional factories, or accelerating their upgrade/enhancement cycle.

Likewise, we can generate a rough Air To Ground ratio, from the Air Threat data, breaking out the A2A versus the A2G information, and deriving a 'bias' value that lets us know if a significant portion of the air power is made up of gunships/bombers.

Again, this can be useful in telling us that we may actually have fighter superiority, even the the basic Air Ratio may say otherwise.   This value is highly indicative of early bomber/gunship rush behavior.