LUA - AirForceAI
The ProsecuteTarget subfunction has been moved out of the behavior into a discrete function.  This became necessary when we decided that we would allow platoon merging while units were actually off doing this.

We'll now carry the current loiterposition on the platoon data, so that if a unit that's in ProsecuteTarget is moved to a new platoon while doing that, it will get the correct loiterposition from the new platoon.