LUA - TML and NUKE firing
With some investigation about the ISSUETACTICAL and ISSUENUKE commands, we find that this seems to bypass certain weapon state activities - and will not clear the target once the firing has taken place.   This has left both TML and NUKE weapons in a state where they no longer fire - but just accumulate additional targets.

We now clear all commands, prior to issuing this command, which clears up the backlog and allows the weapon to fire.