LUA - ScenarioUtilities
Some refactoring, and removal of the threat insertion on each starting location.

Originally, we used the threat insertion to help the AI orient it's start position towards the enemy.   This method has proven to be somewhat unreliable, so we'll simply orient towards map center for now.

As the game develops, bases will orient towards the closest overall threat - so we'll see how that works out.