LOUD_UNITS - Barrage artillery
The complicated gimbals of the barrage artillery get reworked - again - as we try to co-ordinate the turret actions and miniimize the OnFire() discrepancies between the barrels.

This is tricky, since all 3 barrels are gimballed either on the Yaw or the pitch, so some barrels will generate an OnFire() event at a different time than the others.