BO for LOUD - Aeon Purifier
The Purifier is an odd duck in BO Unleashed.   It's intended role as a 'super' TML is often overlooked, and some of it's mechanics tell us why.

The missiles, while splitting, are not target bound, so getting a solid hit is difficult, but as an area saturation weapon, it is useful.

Considered just converting it into a land based version of the tactical missile barrage that is found on the naval Missile Cruiser, but we'll try to breathe some life into this unit before I go down that road.