TM - BRPAT2FIGBO
The weapon on this unit is just rigged poorly - and trying to fire 3 times per second.   The number of projectiles, per rack (2), and the two (2) racks, means the unit took 4 ticks to do one firing - so it was always missing every 2nd fire event - and thus DPS would never line up with what was indicated in the blueprint.

Eliminated the 2nd project - so each rack fires only one - which removes the delay between racks - and I set the Rof down to 2 (every 5 ticks), so the weapon can reliably fire every time.

Damage per projectile goes from 15 to 45 - so that the listed 180 DPS can be obtained.

This does NOT address the aiming issues on this unit, which tries to aim both wingtip weapons, from a 3rd bone on the central line - this - in combination with a Firing tolerance value of 2, results in some missing issues, especially when the unit frame is not lined up with the target, and cannot be properly resolved with splitting the single weapon into two self-aimed weapons, which I'm not about to spend time doing.    This is a small issue.