LUA - Blocking Terrain Check
Standardized the check to an 8 ogrid measure (some were 6) and normalized the maximum deviation to 3.6 ( 40.5 degrees ).

For pathfinding (which covers engineers), the check must now fail twice, in a single iteration, greatly improving the reliability of the check, in the presence of rough terrain - this greatly improves not only the response of the check, but the reliability - reducing the instances of 'pathfinding stall'.