UNITS - Air performance
This is a major revision to air flight mechanics.  Primarily, it identifies some rather errant air mechanics (such as hovering and/or stopping mid-flight, moving backwards, and/or, extreme turns) and codifies and harmonizes those behaviors across air platforms.

Flight parameters, both across classes, and tiers, are harmonized along identifiable lines, and in accordance with realistic expectation.

The T1 flight parameters should be considered the baseline by which the classes are defined (the exception being gunships) - and progressive changes are made thru the tiers.  More detail about this process can be found on the unit spreadsheet where I've laid out the parameters of the classes, by tier, for more immediate reference.