LUA - AI Engineer tasks
A number of engineer tasks have been tweaked here, primarily in turns of sensitivity to local threat or combat ratio.   The main purpose is to curtail building of certain structures in the absence of a credible threat

Energy building (T2 Pgens primarily) were suffering from a serious overlap issue with building of T3 Pgens.

Added more T2 MassFab options, intended to improve adjacency bonuses, in the MAIN base, for the inner ring factories - and to better account for the presence of the T3 mass fabs.