LUA - MAIN base templates
The layout and sequencing of T3 Pgens has been reviewed in order to take better advantage of sequencing of the mass fabs.

The positions of T2 Mass Fabs, and their sequencing, has been updated to add six additional positions

The sequencing of factory positions has been improved to insure better use of the rear line and tertiary positions when MAIN wants to grow upto or beyond the proscribed natural 12 positions.