LUA - Transport Utilities
Transport utilities will now better evaluate threat positions, and deny transport more reliably, rather than simply going direct to goal when an appropriate path cannot be found.

I have also introduced a variable 'AtGoalDistance' value, into the process, so that the builder data can proscribe an 'unload distance'.   For engineers, this value should be set rather low, as this will allow the transport process to abort much later in the process, should it discover that the intended landing zone is no longer safe.

If not defined in the builder, the currently existing value of 10000 (100 ogrids distance) will continue to apply.