LUA - AI Engineer Mass & Energy builders
Tightening up the threat parameters, and the build allowances - this should minimize the T2/T3 overlap where T2 Pgens might get constructed when a T3 Pgen was already underway.

Mass builders now better match the threat tolerances of the condition, with the builder data - this will reduce or eliminate a certain amount of engineer 'thrashing' where the engineer would take a job, and then immediately abandon it due to threat.