LUA - Transportutilities
A number of edge condition fixes here, related mostly to units that got stuck on an airpad - and being assigned either to platoons, or given transport tasks (UEF T2 Gunship).

This has all kinds of follow-on issues, so it's an important correction, dealing with transporting 'immobile' units, or those still attached.   One of the most obvious issues was long range 'warping' of units - which was caused when a stuck transport, often a significant distance away, would 'warp' units to it's position.   Bad - very bad.

The warping was meant to deal with heavy congestion issues, at very short ranges, and that it does very well - but not this.

Added a feature to allow engineers to be less tolerant of loading delays - as this turned out to be a case of an engineer actually arriving, and starting a job, prior to the transport arriving - resulting in a very lengthy wait for the transport.