BO for LOUD - High regen units
This work involves those units with extreme regen rates, in this example, Basilisk.   The goal is to arrive at a fundamental relationship for how 'mass packing' (putting excess mass into a unit) affects not only the base regen rate, but the regen rate bonuses given when gaining veteran level.

Our current formula dictates a 'base' regen rate based upon the gross mass that goes into a unit.    All units use this, and it's shown on the unit spreadsheet.

Any difference between the 'base' regen rate, and the blueprint regen rate, is shown on the spreadsheet as 'excess mass' required to obtain this higher regen.

What I've changed --

Basi's elevated regen goes from 125 >> 100, this is still, by far, the most elevated regen in the game, but this value is much more aligned with how much additional mass goes into the unit, to gain this effect - and still keep Basi combat effective.    Without this change, the firepower/cost ratios were giving Basi a pretty overwhelming 'free pass' in that area.

Regen, earned from veteran levels, is now linked to the base regen, and not the elevated regen rate - so, in this example, Basi has 100 elevated regen, but a base regen of only 15.

The first two vet levels will grant a slightly elevated regen bonus of 20, dropping to 15 for levels 3 and 4, and then 10 for level 5.