LUA - Fog of War timing
When FOW is turned off, the ParseIntel thread can become overloaded, as it now has to process every IMAP block on every pass.   This can, from time to time, cause a crash.

When FOW is turned off, we'll lengthen the ParseIntel process, by an additional 3 seconds, reducing the stress of the additional load.