LUA - Self-Upgrade/Unit Restrictions
The self-upgrade process was not processing unit restrictions, so when restrictions such as Tier2 or Tier3 are used, the process would get stuck repetitively trying, and failing - and essentially bringing about a near complete halt to production.

The process now correctly identifies restrictions, and signals that the self-upgrade path for that unit is complete.

This should permit games with such restrictions to at least play better - but not ideally, since the overall baseplan is not geared for maximum production at lower tiers like that.