BrewLAN for LOUD - T3 Mobile Radar
Harmonized all radar ranges and power consumptions with radar formula.

A general reduction in both range, and power consumption for these units, which is in keeping with reasonable limits on energy consumption for mobile units (that don't have the capacity or size to generate their own).

Reviewed various animations in order to generate more 'action' when the unit is on, versus when it is not.