UNITS - T1 Frigates - Depth Charges
This introduces the very basic depth charge to all T1 frigates.

Depth charges, quite similar to torpedoes, are very short range - and very slow - so the frigates must close to use them - and while they are area effect weapons, the damage is fairly low.

This effectively gives the frigates a limited ability to counter submarines, especially during the early stage of the game, where submarines can be almost unstoppable.

The frigates have a new behavior, which allows them to 'boom and zoom', dashing in to drop the charges, and then zooming out while reloading - prior to coming back in to prosecute the attack.

This behavior is achieved by toggling the minimum range of the weapon, just after firing, to a distance greater than the normal range of the weapon.   When almost reloaded, the range is reset, and the unit will then move to get back into range to fire again.

We can extend this behavior further, by lowering the maximum range, once a target has been acquired, forcing the unit to dash right into a point almost directly overhead of the target, virtually insuring a hit.