LUA - Unpacking Weapon
There's an edge condition with unpacking weapons that can leave the weapon in a non-functional state.   During the unpack sequence, the weapon is being disabled, and what can happen is that either the weapon gets trapped in a loop, returning to idle state (and not firing) - or - the weapon gets 'stuck' in the RackSalvoFireReady() state.

Easily solved, and logical, there's no need to disable a weapon when unpacking - disabling a weapon DOES happen when packing though.