LUA - AI functions - ThreatFuther, ThreatCloser
Both of these functions are revised to be sensitive to both the current LandRatio, and the basic starting cheat value of the AI.   This will make both of these functions, which control several AI tasks, dynamic within the conditions of the game - and his current status.

The cheating impact is fairly minor, but the LandRatio will greatly influence these functions.

Added an EBC version of one of them, which is useful for some of the immediate game start conditions, allow me to reduce the size of the BuildConditions table (always good for performance).