LUA - AI Antinuke builders at Expansion
Was noting some hesitancy for the AI to build Nuke defense at his expansion base - but also a condition where he would build one when it wasn't needed.

I restructured several conditions to make that more responsive to the situation and a little less restrictive.

I increased the acceptable threat maximums on some of the DP engineering tasks, and added the SACU to the Airpad builder