LUA - Ability buttons, loaded save games
Discovered an issue with 'ability' buttons not being restored when you load a save.   For the 'scry' units, since the ability button is the only action they have - they were effectively 'locked out' of the game.

I've added code to detect that situation, and re-initialize the necessary data for them to continue normally.