UNITS - T2 Gunships
This is a rework of firing cycles, to demonstrate some techniques to reduce memory usage - and to make use of the various timing parameters of weapons to create varied firing patterns.

The net effect on the gunships is nil - with DPS remaining the same, but overall, the gunships do this over a 2 second firing cycle, resulting in less projectiles, less processing and less memory usage.

T2 gunships have also had their lift factors revised to be much closer to their T3 brethern and fuel recharge (when landed on the ground) has been reduced.

UAA0203 - demonstrates using a single muzzle bone, rather then 4, to perform the firing cycle.   Also, removed a bunch of nonsense in the script that was obviously meant for something else.

UEA0203 - Since the flash, and the audio, infer a gatling like weapon, you'd think it would have multiple projetiles - it doesn't - just one, but now much larger fired over a 2 second burst.  The flashes are now more properly sized.

URA0203 - this demonstrates using two racks, rather than the previous single rack that fired from two different hard points - and making use of 'RackFireTogether' - along with a rack salvo charge time & a rack salvo reload time - to keep the firing cycle tightly spaced.

XSA0203 - this demonstrates taking what was a single weapon, and breaking it into two (which is the way the model presents it) - and giving slightly different cycle timings to create an interesting firing pattern between the two - using MuzzleSalvoDelay & RackSalvoReloadTime to keep in the 2 second firing cycle.