LUA - AI AirRatio & NavalRatio calcs
Altered this to better resolve early game conditions where the enemy strength has not yet been detected.

For AirRatio, this originally kept the AirRatio suppressed at 0.2 until the enemy air strength was actually detected.

Now, it will climb in relation to .25 times the Air Production ratio, which is based upon the potential factory output.   This will allow the Air ratio to move upwards, even when enemy air units are not detected, assuming that greater production also infers air superiority.

The net effect of this change is allowing the AI to possibly transition to transports, gunships and bombers, at a much earlier stage of the game.