LUA - UI - Fuel Time
In further testing, the new Fuel Time buff was working properly - but - the UI was not displaying it correctly.   The UI relied entirely upon the original blueprint value, and not the actual value.

We now record the fuel time of all units as a stat (FUELTIME) so that we can recall it, when needed, in the UI.